+++ Lewee.sln
@@ -54,8 +54,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sample.Restaurant.Server.Te
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lewee.IntegrationTests", "src\Lewee.IntegrationTests\Lewee.IntegrationTests.csproj", "{27CE2C4C-ED90-44DA-B3E3-E798459E4B5B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lewee.Infrastructure.Logging", "src\Lewee.Infrastructure.Logging\Lewee.Infrastructure.Logging.csproj", "{384767CA-2EEB-4C35-B8C4-400044CDCCB9}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -142,10 +140,6 @@ Global
 		{27CE2C4C-ED90-44DA-B3E3-E798459E4B5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{27CE2C4C-ED90-44DA-B3E3-E798459E4B5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{27CE2C4C-ED90-44DA-B3E3-E798459E4B5B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{384767CA-2EEB-4C35-B8C4-400044CDCCB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{384767CA-2EEB-4C35-B8C4-400044CDCCB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{384767CA-2EEB-4C35-B8C4-400044CDCCB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{384767CA-2EEB-4C35-B8C4-400044CDCCB9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
