+++ sample/Sample.Restaurant.Server/Sample.Restaurant.Server.csproj
@@ -24,7 +24,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Lewee.Infrastructure.AspNet.WebApi\Lewee.Infrastructure.AspNet.WebApi.csproj" />
-    <ProjectReference Include="..\..\src\Lewee.Infrastructure.Logging\Lewee.Infrastructure.Logging.csproj" />
     <ProjectReference Include="..\Sample.Restaurant.Client\Sample.Restaurant.Client.csproj" />
     <ProjectReference Include="..\Sample.Restaurant.Infrastructure\Sample.Restaurant.Infrastructure.csproj" />
   </ItemGroup>
