+++ sample/Sample.Restaurant.Server/appsettings.json
@@ -1,9 +1,4 @@
 {
-  "ApplicationSettings": {
-    "Name": "Sample.Restaurant.Server",
-    "Environment": "Local",
-    "Version": "0.0.1"
-  },
   "SeqSettings": {
     "IsEnabled": true,
     "Uri": "http://seq",
@@ -12,10 +7,5 @@
   },
   "ConnectionStrings": {
     "Sample.Restaurant": "Server=lewee_sql;Database=Lewee.Sample.Restaurant;User=sa;Password=<YourStrong@Passw0rd!>;MultipleActiveResultSets=true;Encrypt=false"
-  },
-  "MigrateDatabases": true,
-  "ServiceBusSettings": {
-    "BusType": "RabbitMQ",
-    "ConnectionStringOrHost": "localhost"
   }
 }
