+++ src/Lewee.Infrastructure.AspNet/Lewee.Infrastructure.AspNet.csproj
@@ -7,6 +7,13 @@
   <ItemGroup>
     <PackageReference Include="Correlate.AspNetCore" />
     <PackageReference Include="Serilog" />
+    <PackageReference Include="Serilog.Enrichers.Environment" />
+    <PackageReference Include="Serilog.Enrichers.Process" />
+    <PackageReference Include="Serilog.Enrichers.Thread" />
+    <PackageReference Include="Serilog.Extensions.Logging" />
+    <PackageReference Include="Serilog.Sinks.Console" />
+    <PackageReference Include="Serilog.Sinks.Seq" />
+    <PackageReference Include="Serilog.Sinks.Trace" />
     <PackageReference Include="SerilogAnalyzer">
       <ExcludeAssets>all</ExcludeAssets>
     </PackageReference>
