+++ src/Lewee.Infrastructure.AspNet/Observability/SeqSettings.cs
