+++ src/Lewee.Infrastructure.AspNet/Observability/SettingsExtensions.cs
