+++ src/Lewee.Infrastructure.AspNet/Observability/VersionHelper.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+
+namespace Lewee.Infrastructure.AspNet.Observability;
+
+internal static class VersionHelper
+{
+    public static string GetVersion()
+    {
+        return Assembly.GetEntryAssembly()?.GetName().Version?.ToString() ?? "1.0.0";
+    }
+}
