--- src/Lewee.Infrastructure.Logging/ApplicationSettings.cs
@@ -1,8 +1,0 @@
-﻿namespace Lewee.Infrastructure.Logging;
-
-internal class ApplicationSettings
-{
-    public string Name { get; set; } = string.Empty;
-    public string Environment { get; set; } = string.Empty;
-    public string Version { get; set; } = string.Empty;
-}
