--- src/Lewee.Infrastructure.Logging/README.md
@@ -1,2 +1,0 @@
-# Lewee.Infrastructure.Logging
-
